refactor(helper): refactored json classes to the support namespace
Refs #47
