+++ src/Support/Json/Decode.php
@@ -11,21 +11,22 @@
 
 declare(strict_types=1);
 
-namespace Phalcon\Helper;
+namespace Phalcon\Support\Json;
 
 use InvalidArgumentException;
 
 use function json_decode;
-use function json_encode;
 use function json_last_error;
 use function json_last_error_msg;
 
 use const JSON_ERROR_NONE;
 
 /**
- * This class offers a wrapper for JSON methods to serialize and unserialize
+ * Class Decode
+ *
+ * @package Phalcon\Support\Json
  */
-class Json
+class Decode
 {
     /**
      * Decodes a string using `json_decode` and throws an exception if the
@@ -53,7 +54,7 @@ class Json
      * @throws InvalidArgumentException if the JSON cannot be decoded.
      * @link http://www.php.net/manual/en/function.json-decode.php
      */
-    final public static function decode(
+    public function __invoke(
         string $data,
         bool $associative = false,
         int $depth = 512,
@@ -68,46 +69,5 @@ class Json
         }
 
         return $decoded;
-    }
-
-    /**
-     * Encodes a string using `json_encode` and throws an exception if the
-     * JSON data cannot be encoded
-     *
-     * ```php
-     * use Phalcon\Helper\Json;
-     *
-     * $data = [
-     *     'one' => 'two',
-     *     'three'
-     * ];
-     *
-     * echo Json::encode($data);
-     * // {"one":"two","0":"three"}
-     * ```
-     *
-     * @param mixed $data    JSON data to parse
-     * @param int   $options Bitmask of JSON decode options.
-     * @param int   $depth   Recursion depth.
-     *
-     * @return string
-     *
-     * @throws InvalidArgumentException if the JSON cannot be encoded.
-     * @link http://www.php.net/manual/en/function.json-encode.php
-     */
-    final public static function encode(
-        $data,
-        int $options = 0,
-        int $depth = 512
-    ): string {
-        $encoded = json_encode($data, $options, $depth);
-
-        if (JSON_ERROR_NONE !== json_last_error()) {
-            throw new InvalidArgumentException(
-                'json_encode error: ' . json_last_error_msg()
-            );
-        }
-
-        return (string) $encoded;
     }
 }
